Persist tomcat configuration
Do not install tomcat configuration files with each run of the
role to ensure that manual updates to the files are not overwritten.
(Not all aspects of tomcat are managed with this role right now.)

Further ensure that unconfigured tomcat ports are addressed with
default values using the jinja |default() function.